+++ app/routes/components/Input/changelog.codegen.ts
@@ -10,4 +10,7 @@ export const changelog = ''
   + '  adjust height and font size. Etc.\n'
   + '- 2025-09-26: Use theme tokens for Input component styles.\n'
   + '- 2025-09-30: Use `InputBase` as a wrapper for the Input component.\n'
+  + '- 2025-10-05: Support the `leftSection` and `rightSection` props for Input\n'
+  + '  component to add elements inside the input field on the left or right side.\n'
+  + '  Adjust padding based on presence of these sections.\n'
   ;