+++ app/routes/components/Input/index.tsx
@@ -1,15 +1,14 @@
+import { ComponentTemplate } from "@/components/ComponentTemplate";
+import { Section } from "@/components/Section";
+
 import type { Route } from "./+types/index";
 
-import { ComponentTemplate } from "src/components/ComponentTemplate";
-import { Section } from "src/components/Section";
-
-import { sourceCode } from "./source_code.codegen";
-import { demoSourceCode } from "./Demo_source_code.codegen";
 import { changelog } from "./changelog.codegen";
-
 import { Demo } from "./Demo";
+import { demoSourceCode } from "./Demo_source_code.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "Input | Copy UI" },
     { name: "description", content: "The Input component from Copy UI." },
@@ -19,12 +18,11 @@ export function meta({}: Route.MetaArgs) {
 export default function InputPage() {
   return (
     <ComponentTemplate component="Input">
-      <Section
-        title="Input"
-        demoAndCode={[<Demo />, demoSourceCode]}
-        sourceCode={sourceCode}
-        changelog={changelog}
-      />
+      <Section.Root title="Input">
+        <Section.Demo node={<Demo />} code={demoSourceCode} />
+        <Section.SourceCode files={sourceCode} />
+        <Section.Changelog changelog={changelog} />
+      </Section.Root>
     </ComponentTemplate>
   );
 }