+++ src/components/Input/CHANGELOG.md
@@ -6,3 +6,6 @@
   adjust height and font size. Etc.
 - 2025-09-26: Use theme tokens for Input component styles.
 - 2025-09-30: Use `InputBase` as a wrapper for the Input component.
+- 2025-10-05: Support the `leftSection` and `rightSection` props for Input
+  component to add elements inside the input field on the left or right side.
+  Adjust padding based on presence of these sections.